BASW-175: Refactor Membership Pre Hook into Separate Classes
Refactored the class that implemented the Pre hook for memberships into two
separate classes, on for 'create', the other one for 'edit' ops. Also moved
the code that was on the PreEdit_Membership class into the new
Pre_MembershipEdit class.
