+++ membershipextras.php
@@ -161,8 +161,13 @@ function membershipextras_civicrm_pre($op, $objectName, $id, &$params) {
     $contributionID = $id;
   }
 
-  if ($objectName === 'Membership') {
-    $membershipPreHook = new CRM_MembershipExtras_Hook_Pre_Membership($op, $id, $params, $contributionID);
+  if ($objectName === 'Membership' && $op == 'create') {
+    $membershipPreHook = new CRM_MembershipExtras_Hook_Pre_MembershipCreate($params);
+    $membershipPreHook->preProcess();
+  }
+
+  if ($objectName === 'Membership' && $op == 'edit') {
+    $membershipPreHook = new CRM_MembershipExtras_Hook_Pre_MembershipEdit($id, $params, $contributionID);
     $membershipPreHook->preProcess();
   }
